+++ modules/SACT/sprite_eupdate.c
@@ -59,7 +59,7 @@ int sp_eupdate(int type, int time, int cancel) {
 	int edtime = curtime + time * 10;
 
 	while ((curtime = sdl_getTicks()) < edtime) {
-		sdl_effect_step(eff, (double)(curtime - sttime) / (edtime - sttime));
+		sdl_effect_step(eff, (float)(curtime - sttime) / (edtime - sttime));
 		int rest = 16 - (sdl_getTicks() - curtime);
 		int key = sys_keywait(rest, cancel ? KEYWAIT_CANCELABLE : KEYWAIT_NONCANCELABLE);
 		if (cancel && key)
